Refactored Todo feature.
